+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Added v42 OpenAPI spec and generated API.
+- Added v42 client library support.
 
 ### Fixed